Pull script now grabs almost all info from all courses on catalogue with out exceptions
