+++ updateLog.txt
@@ -1,2 +1,3 @@
 01/08/2024 Removed trigger logic involved with creating records upon a seat becoming available within the Database-KG
 01/08/2024 Removed status column from the alerts table.
+01/08/2024 Pull script now grabs almost all info from all courses on catalogue without exceptions-ILC
